STM32 Nucleo F746ZG Hal Drivers
Copied over HAL driver sfor Nucleo F746ZG into approrpirate directors in Motate project structure. These are the base driver files for the chip as it is on the Nucleo board we are using for prototype and development. The next step is to create the Motate specific files that link these drivers to the abstraction layer and then application layer.